+++ snowpack/Snowpack.h
@@ -128,7 +128,7 @@ class Snowpack {
 		void setHydrometeorMicrostructure(const CurrentMeteo& Mdata, const bool& is_surface_hoar, ElementData &EMS);
 
 		void fillNewSnowElement(const CurrentMeteo& Mdata, const double& length, const double& density,
-		                        const bool& is_surface_hoar, const unsigned int& number_of_solutes, ElementData &elem);
+		                        const bool& is_surface_hoar, const unsigned short& number_of_solutes, ElementData &elem);
 
 		void compSnowFall(const CurrentMeteo& Mdata, SnowStation& Xdata, double& cumu_hnw,
 		                  SurfaceFluxes& Sdata);
